[feature]添加转场动画

+++ Demo/libs/packaging/function/encryptionAndDecryption/RSA/ZYQRSA.m
@@ -130,7 +130,7 @@ static NSData *base64_decode(NSString *str){
                                             outbuf,
                                             &outlen);
         if (status != 0) {
-            NSLog(@"SecKeyEncrypt fail. Error Code: %d", status);
+            NSLog(@"SecKeyEncrypt fail. Error Code: %d", (int)status);
             ret = nil;
             break;
         }else{
@@ -325,7 +325,7 @@ static NSData *base64_decode(NSString *str){
                                &outlen
                                );
         if (status != 0) {
-            NSLog(@"SecKeyEncrypt fail. Error Code: %d", status);
+            NSLog(@"SecKeyEncrypt fail. Error Code: %d", (int)status);
             ret = nil;
             break;
         }else{
@@ -409,7 +409,7 @@ static NSData *base64_decode(NSString *str){
                                 cipherBuffer,
                                 &cipherBufferSize
                                 );
-    NSAssert(sanityCheck == noErr, @"加密错误，OSStatus == %d", sanityCheck);
+    NSAssert(sanityCheck == noErr, @"加密错误，OSStatus == %d", (int)sanityCheck);
     // 生成密文数据
     cipher = [NSData dataWithBytes:(const void *)cipherBuffer length:(NSUInteger)cipherBufferSize];
     if (cipherBuffer) free(cipherBuffer);
