+++ Demo/module/function/controller/search/ZYQSearchController.m
@@ -72,7 +72,7 @@
     if (!cell) {
         cell = [[UITableViewCell alloc] initWithStyle:UITableViewCellStyleValue1 reuseIdentifier:@"cell"];
     }
-    cell.textLabel.text = [NSString stringWithFormat:@"第几行%ld" , indexPath.row];
+    cell.textLabel.text = [NSString stringWithFormat:@"第几行%ld" , (long)indexPath.row];
     
     return cell;
 }
